removed AddDescAuto component, autosuggest for descriptions keystroke behind bug fixed
