+++ front/app/components/FacetCard/AddFacetCard.tsx
@@ -1,13 +1,12 @@
 import * as React from 'react';
 import styled from 'styled-components';
 import withTheme from 'containers/ThemeProvider';
-import { FormControl } from 'react-bootstrap';
-import * as FontAwesome from 'react-fontawesome';
 import ThemedButton from 'components/StyledComponents';
 import * as Autosuggest from 'react-autosuggest';
 import AddFieldAuto from 'components/FacetCard/AddFieldAuto';
-import AddDescAuto from 'components/FacetCard/AddDescAuto';
 import ThemedAutosuggestButton from 'components/StyledComponents';
+import { ApolloConsumer } from 'react-apollo';
+import { gql } from 'apollo-boost';
 
 const MarginContainer = styled.div`
   margin: 4px;
@@ -30,6 +29,49 @@ const Row = styled.div`
 
 const ThemedCenterButton = withTheme(CenterButton);
 
+const AUTOSUGGEST_QUERY = gql`
+  query CrumbsSearchPageAggBucketsQuery(
+    $agg: String!
+    $q: SearchQueryInput!
+    $aggFilters: [AggFilterInput!]
+    $crowdAggFilters: [AggFilterInput!]
+    $page: Int!
+    $pageSize: Int!
+    $aggOptionsFilter: String
+    $aggFields: [String!]!
+    $crowdAggFields: [String!]!
+    $url: String
+  ) {
+    autocomplete(
+      aggFields: $aggFields
+      crowdAggFields: $crowdAggFields
+      url: $url
+      params: {
+        agg: $agg
+        q: $q
+        sorts: []
+        aggFilters: $aggFilters
+        crowdAggFilters: $crowdAggFilters
+        aggOptionsFilter: $aggOptionsFilter
+        aggOptionsSort: [{ id: "count", desc: true }]
+        page: $page
+        pageSize: $pageSize
+      }
+    ) {
+      autocomplete {
+        name
+        isCrowd
+        results {
+          key
+          docCount
+        }
+        __typename
+      }
+      __typename
+    }
+  }
+`;
+
 interface AddFacetCardProps {
   submitFacet?: any;
   upsert?: any;
@@ -39,6 +81,7 @@ interface AddFacetCardProps {
   aggNames?: any;
   siteView?: any;
   values?: any;
+  showAddFacet: boolean;
 }
 
 interface AddFacetCardState {
@@ -46,6 +89,8 @@ interface AddFacetCardState {
   description: string;
   showForm: boolean;
   descDisabled: boolean;
+  descriptionSuggestions: any[];
+  isSuggestionLoading: boolean;
 }
 
 class AddFacetCard extends React.PureComponent<
@@ -57,6 +102,8 @@ class AddFacetCard extends React.PureComponent<
     descDisabled: true,
     title: '',
     description: '',
+    descriptionSuggestions: [],
+    isSuggestionLoading: false,
   };
 
   handleButtonClick = user => {
@@ -75,9 +122,67 @@ class AddFacetCard extends React.PureComponent<
     });
   };
 
-  handleDescriptionFieldChange = (e, { newValue }) => {
+  handleDescriptionFieldChange = (e, { newValue }, apolloClient) => {
+    console.log('desc', newValue);
+    this.setState(
+      {
+        description: newValue,
+      },
+      () => {
+        this.getSuggestions(apolloClient);
+      }
+    );
+  };
+
+  getSuggestions = async apolloClient => {
+    const { values } = this.props;
+    const { title, description } = this.state;
+    const query = AUTOSUGGEST_QUERY;
+    const variables = {
+      agg: 'browse_condition_mesh_terms',
+      aggFilters: [],
+      aggOptionsFilter: description,
+      crowdAggFilters: [],
+      page: 0,
+      pageSize: 5,
+      q: {
+        children: [],
+        key: 'AND',
+      },
+      sorts: [],
+      aggFields: [],
+      crowdAggFields: [title],
+    };
+    const response = await apolloClient.query({
+      query,
+      variables,
+    });
+
+    const array = response.data.autocomplete.autocomplete[0].results;
+
+    array.map(({ key, docCount }, i) => {
+      values?.map(([value, checked]) => {
+        if (key === value) {
+          if (checked) {
+            array.splice(i, 1);
+          }
+          if (key === '-99999999999') {
+            array.splice(i, 1);
+          } else {
+            // console.log('good data', array[i]);
+          }
+        }
+      });
+    });
+
+    const suggestions = [
+      { key: description.trim(), partialString: true },
+      ...array,
+    ];
+
     this.setState({
-      description: newValue,
+      descriptionSuggestions: suggestions,
+      isSuggestionLoading: false,
     });
   };
 
@@ -106,10 +211,54 @@ class AddFacetCard extends React.PureComponent<
     });
   };
 
+  onSuggestionSelected = (
+    event,
+    { suggestion, suggestionValue, suggestionIndex, method }
+  ) => {
+    this.setState({
+      description: suggestionValue,
+    });
+  };
+
+  onSuggestionsFetchRequested = async apolloClient => {
+    this.setState({
+      isSuggestionLoading: true,
+    });
+  };
+
+  onSuggestionsClearRequested = () => {
+    this.setState({
+      descriptionSuggestions: [],
+      isSuggestionLoading: true,
+    });
+  };
+
+  getSuggestionValue = suggestion => {
+    return suggestion.key;
+  };
+
+  renderFieldSuggestion = suggestion => {
+    // const capitalized = capitalize(suggestion);
+    if (suggestion.partialString) {
+      return (
+        <Row>
+          <span>{`add "${suggestion.key}" as new description`}</span>
+          <ThemedAutosuggestButton>Add</ThemedAutosuggestButton>
+        </Row>
+      );
+    }
+    return (
+      <Row>
+        <span>{`${suggestion.key}`}</span>
+        <ThemedAutosuggestButton>Add</ThemedAutosuggestButton>
+      </Row>
+    );
+  };
+
   render() {
-    const { showForm, title, description } = this.state;
-    const { user, aggNames, siteView, values } = this.props;
-    if (showForm) {
+    const { title, description, descriptionSuggestions } = this.state;
+    const { aggNames, showAddFacet } = this.props;
+    if (showAddFacet) {
       return (
         <MarginContainer>
           <MarginContainer>Label</MarginContainer>
@@ -120,13 +269,26 @@ class AddFacetCard extends React.PureComponent<
             onSuggestionSelected={this.onFieldSuggestionSelected}
           />
           <MarginContainer>Description</MarginContainer>
-          <AddDescAuto
-            siteView={siteView}
-            values={values}
-            handleInputChange={this.handleDescriptionFieldChange}
-            description={description}
-            title={title}
-          />
+          <ApolloConsumer>
+            {apolloClient => (
+              <Autosuggest
+                suggestions={descriptionSuggestions}
+                renderSuggestion={this.renderFieldSuggestion}
+                inputProps={{
+                  disabled: title ? false : true,
+                  value: description ? description : '',
+                  onChange: (e, value) =>
+                    this.handleDescriptionFieldChange(e, value, apolloClient),
+                }}
+                onSuggestionSelected={this.onSuggestionSelected}
+                onSuggestionsFetchRequested={() =>
+                  this.onSuggestionsFetchRequested(apolloClient)
+                }
+                onSuggestionsClearRequested={this.onSuggestionsClearRequested}
+                getSuggestionValue={this.getSuggestionValue}
+              />
+            )}
+          </ApolloConsumer>
           <div style={{ marginTop: 5, marginLeft: 2, marginBottom: 5 }}>
             <ThemedButton
               style={{ marginRight: 5 }}
@@ -138,16 +300,7 @@ class AddFacetCard extends React.PureComponent<
         </MarginContainer>
       );
     }
-    return (
-      <ThemedCenterButton>
-        <FontAwesome
-          style={{ color: 'inherit', fontSize: 'inherit', margin: 'auto' }}
-          inverse={true}
-          name="plus-square-o"
-          onClick={() => this.handleButtonClick(user)}
-        />
-      </ThemedCenterButton>
-    );
+    return null;
   }
 }
