+++ front/app/components/StyledComponents/index.ts
@@ -105,7 +105,6 @@ const PresearchCard = styled.div`
   border-color: ${props => props.theme.buttonSecondary};
   margin: 10px;
   flex: 1;
-  min-height: 200px;
   min-width: 320px;
   max-width: 320px;
   background: white;
@@ -232,4 +231,4 @@ export const StyledProfileForm = styled(FormControl)`
 export const StyledInterventionHeading = styled.div`
   margin: 1em;
   border-bottom: solid white;
-`
+`;
